Moved cypress component tests into each components folder
