+++ web/components/Footer/Footer.cy.js
@@ -1,4 +1,4 @@
-import { Footer } from '../../components/Footer';
+import { Footer } from '.';
 
 const localData = {
   footerLinks: [
