+++ web/components/MetaData/MetaData.cy.js
@@ -1,4 +1,4 @@
-import { MetaData } from '../../components/MetaData';
+import { MetaData } from '.';
 
 const localData = {
   document: {
