+++ web/cypress.config.ts
@@ -6,11 +6,11 @@ export default defineConfig({
       // implement node event listeners here
     }
   },
-
   component: {
     devServer: {
       framework: 'next',
       bundler: 'webpack'
-    }
+    },
+    specPattern: ['./cypress/component/*.cy.js', './components/*/*.cy.js']
   }
 });
